+++ src/frontend/resolver.zig
@@ -144,7 +144,7 @@ const Resolver = struct {
                 try self.resolveNode(catch_clause.body, catch_scope);
             },
             .FinallyClause => |finally_clause| try self.resolveNode(finally_clause.body, scope),
-            .BreakStatement, .ContinueStatement => {},
+            .BreakStatement, .ContinueStatement, .DebuggerStatement => {},
             .LabeledStatement => |labeled| try self.resolveNode(labeled.body, scope),
             .ExpressionStatement => |statement| try self.resolveNode(statement.expression, scope),
             .Identifier => |identifier| try self.addReference(node_id, identifier.name, scope, .read),